split up server.ts
